Refactor helpers and migration code
