+++ end-to-end/tests/roles.js
@@ -3,7 +3,7 @@ import { Role } from 'testcafe';
 /**
  * Drupal administrator via local login
  */
-export const localAdmin = Role('https://islandora-idc.traefik.me/user/login', async t => {
+export const localAdmin = Role('https://islandora-idc.traefik.me/user/login', async (t) => {
   await t
     .typeText('#edit-name', 'admin')
     .typeText('#edit-pass', 'password')
